Extract banned_external_code linter to new file
